Add FALLTHROUGH_A, UNUSED_A, fix ssize_t use, tighten tmp path checks
- Introduce FALLTHROUGH_A and UNUSED_A macros in internal/features.h
  for portable attribute annotation (fallthrough and unused)
  across compilers and C standards.
  Update code to use these for clarity and static analysis.
- Replace legacy C "/* FALLTHROUGH */" comments
  with the attribute macro FALLTHROUGH_A in all btree
  and recno switch/case blocks
  (e.g., bt_delete.c, bt_put.c, bt_seq.c, rec_seq.c)
  for correct compiler handling and warning suppression.
- Add UNUSED_A macros to function parameters in db/db.c, mpool_open,
  and hash_open for unused arguments to prevent warnings,
  and apply consistent void casts.
- Ensure all IO syscalls (write, writev, pwrite, pread) use ssize_t
  and check for SSIZE_MAX overflows. Replace int usage with ssize_t
  where necessary and add length checks before IO. Improves safety
  and correctness on 32/64-bit platforms.
- Fix path length checks for tmp file creation
  to ensure snprintf casting/comparison is safe
  (bt_open.c, hash_page.c, ndbm.c), preventing buffer overflows
  and aligning with "int" type signatures.
- Defensive initialization: set local stack variables
  (e.g. EPG *e = NULL, indx_t idx = 0)
  rather than leaving them uninitialized.
- hash.c: validate critical values (H_CHARKEY, HIGH_MASK, etc.)
  for sanity before use (no negative/zero).
- Defensive checks on pointer returns: catch NULL before dereference.
- Improves code compliance for modern static analyzers
  and hardens portability/safety,
  supporting compilation on GCC, Clang, musl, glibc, all BSDs,
  and with strict CFLAGS. No functional logic changes.

+++ btree/bt_open.c
@@ -390,8 +390,8 @@ tmp(void)
 	if (issetugid() == 0)
 		envtmp = getenv("TMPDIR");
 	len = snprintf(path,
-	    sizeof(path), "%s/bt.XXXXXX", envtmp ? envtmp : "/tmp");
-	if (len < 0 || len >= sizeof(path)) {
+	    sizeof path, "%s/bt.XXXXXX", envtmp ? envtmp : "/tmp");
+	if (len < 0 || len >= (int)sizeof path) {
 		errno = ENAMETOOLONG;
 		return -1;
 	}
